Mindre forbedringer på css og refaktorering.
