+++ src/frontend/komponenter/Fagsak/Vedtak/BegrunnelserTabell/begrunnelser-tabell.less
@@ -1,10 +1,13 @@
-.begrunnelser-tabell {
-    padding-bottom: 1rem;
+.vedtaksbrev {
+    & .tabell {
+        margin-bottom: 1rem;
+    }
 }
 
 .begrunnelse-input {
     &__flex {
         display: flex;
+
         .skjemaelement {
             padding: 0 1rem 1rem 0;
         }
